unfucks the team from what ever mess they presented...
