+++ src/App.jsx
@@ -1,33 +1,35 @@
 import React from "react";
 import { connect } from "react-redux";
-import DisplayAllArticles from "./components/DisplayAllArticles";
+import { useEffect } from 'react'
+import DisplayArticles from "./components/DisplayArticles";
 import { fetchArticles } from "./state/actions/articleActions";
 import { bindActionCreators } from "redux";
 import DisplaySingleArticle from "./components/DisplaySingleArticle";
-import DisplayArticlesByCategory from "./components/DisplayArticlesByCategory";
 
 const App = props => {
-  props.fetchArticles();
-  return (
-    <>
-      {props.showArticlesList && <DisplayAllArticles />}
-      {props.singleArticle && <DisplaySingleArticle />}
-      {props.showArticlesByCategory && <DisplayArticlesByCategory />}
-    </>
-  );
+	useEffect(() => {
+		props.fetchArticles();
+	}, [])
+
+	return (
+		<>
+			{props.showArticlesList && <DisplayArticles />}
+			{props.singleArticle && <DisplaySingleArticle />}
+		</>
+	);
 };
 const mapDispatchToProps = dispatch => {
-  return {
-    fetchArticles: bindActionCreators(fetchArticles, dispatch)
-  };
+	return {
+		fetchArticles: bindActionCreators(fetchArticles, dispatch)
+	};
 };
 
 const mapStateToProps = state => {
-  return {
-    showArticlesList: state.showArticlesList,
-    singleArticle: state.singleArticle,
-    showArticlesByCategory: state.showArticlesByCategory
-  };
+	return {
+		showArticlesList: state.showArticlesList,
+		singleArticle: state.singleArticle,
+		showArticlesByCategory: state.showArticlesByCategory
+	};
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(App);
