+++ src/state/actions/articleActions.js
@@ -9,7 +9,6 @@ const fetchArticles = () => {
 };
 
 const fetchSingleArticle = articleID => {
-  debugger
   return async dispatch => {
     let response = await axios.get(`/articles/${articleID}`);
     return dispatch(dispatchSingleArticleAction(response.data));
@@ -21,7 +20,6 @@ const dispatchArticleAction = json => {
 };
 
 const dispatchSingleArticleAction = json => {
-  debugger
   return { type: GET_SINGLE_ARTICLE, payload: json.article };
 };
